move test_plan logic into contained file
